SLING-10236 : Introduce Operation.asRepoInitString (move AsRepoInitTest such that Operation.getParameterDescription is accessible to the test)
